+++ tests/test_cases.py
@@ -6,9 +6,10 @@
 # PFNET is released under the BSD 2-clause license. #
 #***************************************************#
 
+import os
 import glob
 
-DATA_DIR = './data/'
+DATA_DIR = os.path.join('data','')
 MAT_CASES = glob.glob(DATA_DIR+'*.mat')
 RAW_CASES = glob.glob(DATA_DIR+'*.raw')
 CASES = glob.glob(DATA_DIR+'*')